Rename toast folder
