+++ src/entities/sessions/hooks/useSessions/useSessions.ts
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import axios from "axios";
 import path from "../../../../routers/paths/paths";
 import { SessionsStructure } from "../../types";
-import showToast from "../../../../pata/showToast";
+import showToast from "../../../../toast/showToast";
 
 const apiUrl = import.meta.env.VITE_API_URL;
